make results cache code more readable
